Hotfix

+++ immucan_roi/_immucan_roi.py
@@ -62,10 +62,11 @@ class IMMUcanNappingApplication(NappingApplication):
         transf_coords_dir.mkdir(exist_ok=True)
         empty_source_rois = pd.DataFrame(columns=["Name", "X", "Y", "W", "H"])
         for source_img_file in source_img_dir.glob("*"):
-            source_coords_file_name = source_img_file.with_suffix(".csv").name
-            source_coords_file = source_coords_dir / source_coords_file_name
-            if not source_coords_file.exists():
-                empty_source_rois.to_csv(source_coords_file, index=False)
+            if source_img_file.is_file():
+                source_coords_file_name = source_img_file.with_suffix(".csv").name
+                source_coords_file = source_coords_dir / source_coords_file_name
+                if not source_coords_file.exists():
+                    empty_source_rois.to_csv(source_coords_file, index=False)
         dialog = super(IMMUcanNappingApplication, self)._create_dialog()
         dialog.selection_mode = NappingDialog.SelectionMode.DIR
         dialog.matching_strategy = NappingDialog.MatchingStrategy.ALPHABETICAL
